Amazing Progress Today
WOW. Big day. Worked about 5 hours today on the game and I added/fixed a lot. FINALLY fixed the firefighter shooting direction. Made several different prefabs of the water projectiles so that each fireman had a unique projectile. Tweaked the code for the firecontroller and waterball. Added moral grade and rating to the end of the game. Added quit and restart buttons for when you complete game.  Fixed audio for killed stamp.